Cleanup tests. Add tests on stats
